Flatten decode method
Before this commit, the DjvuRleDecoder.decode method had two nested
decoder functions (one for bitonal and one for color images) and would just
call the corresponding one depending on the image mode. This allowed for
the use, in the nested functions, of local variables instead of
attributes which at that time seemed simpler to me.

However, the method with the nested functions ended up being too long
and the functions handling too many things, so I've decided to convert
those functions to private methods. That required changing some variable
and method names, and moving some parts to the decode method.