Correctly check magic number
The `_accept` method is not always called, for example if the plugin is
used directly, like:
`DjvuRleImagePlugin.DjvuRleImageFile(fp=io.BytesIO(b'...'))`, so `_open`
was modified to check the magic number too. Also, there can be no
comments before or in between the magic number, so a simple method to
read it was added. (See comments in Pillow's discussion #5121.)